Silence alarms and re-activate them given an id indefinetelly  (#44)
* changes in order to be able to silence alarms

* activate alarm and general parameter checking

* some verifications added for parameter parsing

* alarm added messsage now contains id for easier identificaiton

+++ commands/deletealarm.js
@@ -15,35 +15,40 @@ module.exports = {
     description: 'Deletes the alarm with a given id - **THIS ACTION CANNOT BE REVERTED**',
     usage: auth.prefix + 'deleteAlarm <id>',
     async execute(msg, args, client, cron, cron_list, mongoose) {
-        var alarm_to_delete = args[0];
-        if (!can_delete_alarm(msg, alarm_to_delete)) {
-            msg.channel.send(`The alarm you selected isn't yours or you aren't administrator on this server therefore you cannot delete it!`)
-        }
-        else {
-            if (cron_list[alarm_to_delete] !== undefined) {
-                try {
-                    if (alarm_to_delete.includes(private_flag)) {
-                        await Private_alarm_model.deleteOne(
-                            { alarm_id: alarm_to_delete }
-                        )
-                    } else if (!alarm_to_delete.includes(temp_flag)) {
-                        await Alarm_model.deleteOne(
-                            { alarm_id: alarm_to_delete }
-                        )
-                    }
-                    cron_list[alarm_to_delete].stop();
-                    delete cron_list[alarm_to_delete];
-                    msg.channel.send(`Sucessfully deleted alarm: ${alarm_to_delete}.`);
-                } catch (e) {
-                    logging.logger.info(`Error deleting alarm with id:${alarm_to_delete}... Please try again later!`);
-                    logging.logger.error(e);
-                    msg.channel.send(`Error deleting alarm with id:${alarm_to_delete}... Please try again later!`);
-                }
+        if (args.length >= 1) {
+            var alarm_to_delete = args[0];
+            if (!can_delete_alarm(msg, alarm_to_delete)) {
+                msg.channel.send(`The alarm you selected is not yours or you aren't administrator on this server therefore you cannot delete it!`)
             }
             else {
-                var myalarms_command_stg = "`myalarms`";
-                msg.channel.send(`Impossible to delete alarm with id ${alarm_to_delete}.\nPlease check if you entered the id of the alarm correctly!\nTo see your alarms, type: ${auth.prefix}${myalarms_command_stg}`);
+                if (cron_list[alarm_to_delete] !== undefined) {
+                    try {
+                        if (alarm_to_delete.includes(private_flag)) {
+                            await Private_alarm_model.deleteOne(
+                                { alarm_id: alarm_to_delete }
+                            )
+                        } else if (!alarm_to_delete.includes(temp_flag)) {
+                            await Alarm_model.deleteOne(
+                                { alarm_id: alarm_to_delete }
+                            )
+                        }
+                        cron_list[alarm_to_delete].stop();
+                        delete cron_list[alarm_to_delete];
+                        msg.channel.send(`Sucessfully deleted alarm: ${alarm_to_delete}.`);
+                    } catch (e) {
+                        logging.logger.info(`Error deleting alarm with id:${alarm_to_delete}... Please try again later!`);
+                        logging.logger.error(e);
+                        msg.channel.send(`Error deleting alarm with id:${alarm_to_delete}... Please try again later!`);
+                    }
+                }
+                else {
+                    var myalarms_command_stg = "`myalarms`";
+                    msg.channel.send(`Impossible to delete alarm with id ${alarm_to_delete}.\nPlease check if you entered the id of the alarm correctly!\nTo see your alarms, type: ${auth.prefix}${myalarms_command_stg}`);
+                }
             }
+        } else {
+            msg.channel.send(`No arguments were passed to execute this command.\n`
+                + 'Usage: ' + auth.prefix + 'deleteAlarm <id>');
         }
 
     }
